backend/requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv==1.1.1
